Create db and log dirs post sync
(Writing PR comment here for now)
Purpose:
seqreports db and log dir are not included in the sync. Therefore, add
it to the shell script that creates these directories post sync.
(triggered by crontab probably).

It would be nice if we could instead include these directories into the
sync. As of now, these variables are scattered through the repo, and
it's hard to understand the code.